requirements.txt
asgiref==3.3.1
Django==3.1.5
mysqlclient==2.0.3
mysql-connector-python==8.0.23
pytz==2020.5
sqlparse==0.4.1
pandas~=1.2.1
SQLAlchemy~=1.3.22
matplotlib~=3.3.4